move ai overview generation endpoints to admin router
