+++ backend/routes/units.js
@@ -6,7 +6,6 @@ const Unit = require('../models/unit');
 
 // Function Imports
 const { verifyAdmin } = require('../utils/verify_token.js');
-const AiOverviewService = require('../services/aiOverview.service');
 const {
   getSortCriteria,
   requiresReviews,
@@ -409,83 +408,6 @@ router.put('/update/:unitcode', verifyAdmin, async function (req, res) {
     });
   }
 });
-
-/**
- * ! POST Regenerate AI overview for all units
- *
- * @async
- * @param {boolean} force - Regenerate even if stored copy is fresh (default false)
- * @param {number} delayMs - Throttle between requests (default service value)
- */
-router.post('/ai-overview/regenerate', verifyAdmin, async function (req, res) {
-  // #swagger.tags = ['Units']
-  // #swagger.summary = 'Admin-only endpoint to rebuild AI overviews across all units with human reviews'
-
-  try {
-    const { force = false, delayMs } = req.body || {};
-    const result = await AiOverviewService.generateOverviewsForAllUnits({
-      force: Boolean(force),
-      delayMs: typeof delayMs === 'number' ? delayMs : undefined,
-    });
-
-    return res.status(200).json({
-      message: 'AI overviews regeneration completed',
-      result,
-    });
-  } catch (error) {
-    return res
-      .status(500)
-      .json({ error: `Failed to regenerate AI overviews: ${error.message}` });
-  }
-});
-
-/**
- * ! POST Regenerate AI overview for a specific unit
- *
- * @async
- */
-router.post(
-  '/:unitcode/ai-overview/regenerate',
-  verifyAdmin,
-  async function (req, res) {
-    // #swagger.tags = ['Units']
-    // #swagger.summary = 'Admin-only endpoint to rebuild the AI overview for a single unit'
-
-    try {
-      const unitCode = req.params.unitcode.toLowerCase();
-      const { force = true } = req.body || {};
-
-      const unit = await Unit.findOne({ unitCode });
-      if (!unit) return res.status(404).json({ error: 'Unit not found' });
-
-      const result = await AiOverviewService.generateOverviewForUnit(unit, {
-        force: Boolean(force),
-      });
-
-      if (result.status === 'skipped') {
-        return res
-          .status(200)
-          .json({ message: 'No regeneration required', result });
-      }
-      if (result.status === 'updated') {
-        await unit.populate('reviews', '_id');
-        return res.status(200).json({
-          message: 'AI overview updated',
-          overview: unit.aiOverview,
-          result,
-        });
-      }
-
-      return res
-        .status(500)
-        .json({ error: 'Failed to regenerate AI overview', result });
-    } catch (error) {
-      return res
-        .status(500)
-        .json({ error: `Failed to regenerate AI overview: ${error.message}` });
-    }
-  }
-);
 
 /**
  * ! GET Units Required-By
